\name{HMMRel-package}
\alias{HMMRel-package}
\docType{package}
\title{HMMRel - Hidden Markov Models for Reliablity and Maintenance}

\description{
 Reliability Analysis and Maintenance Optimization using Hidden Markov Models (HMM).
}

\details{
  \tabular{ll}{
    \strong{Package:} \tab HMMRel \cr
    \strong{Type:} \tab Package \cr
    \strong{Version:} \tab 1.0 \cr
    \strong{Date:} \tab 2024-11-11 \cr
    \strong{License:} \tab GPL version 2 or later \cr
    \strong{Maintainer:} \tab M.L. Gamiz <mgamiz@ugr.es> \cr
    \strong{URL:} \tab \url{http://wpd.ugr.es/~reliastat} \cr
  }
}

\author{
  M.L. Gamiz, N.Limnios, and M.C. Segovia-Garcia
}
\references{Gamiz, M.L., Limnios, N., Segovia-Garcia, M.C. (2023). Hidden Markov models in reliability and maintenance. European Journal of Operational Research, 304(3), 1242-1255.
}

